Added audio, including switching. Webapp re-order.
This will likely need some work, but audio is properly captured and
encoded. The gst pipelines are awful now, but that's just how gstreamer
like them.

Timestamps on the audio appear to work correctly, there doesn't appear
to be any lag and switching doesn't produce pops or anything weird.

For the webapp, added an audio input toggle, but also re-arranged some
things, moving things that should be hard to bump into collapsible
menus.