settings file now applies to gunicorn.
Also fixed logging to make it so that the new logging setting would be
applied to gunicorn. All of these settings can be overwritten by the
command line switches.

Added time since a thread was started, to help try to track down a bug,
and a cleanup when ctrl+C is pushed, which also prints this.