Refactor of SRT stats gathering thread.
This was mostly to clean things up so that the underlying process could
be cleanly restarted to work around a potential SRT bug related to it
complaining about the time on the sender and receiver being out of sync
somehow.
The error is "Wrong source time was provided.Sending is rejected.", and
only appearred after running for hours (potentially idle, I haven't done
a test leaving it sending for hours to see if the error case happens
then too).