Added full-on IPC to share state.
It's effectively a single bit, because all I'm tracking in whether or
not the scene is locked between threads, because the API workers
couldn't get a clear view of whether it was, or not. This meant that the
client calling /status/ wouldn't get a reasonable result consistently.

More comments inline, but using multiprocessing.Manager to do IPC. Could
also use this to clear up some of the issues with spawning 1x websocket
connection per worker.